adversary: Introduce skeleton framework for adversarial features (#16)
This framework is meant to provide a uniform way to add new adversarial
features, configure those adversarial features, and to read and adjust
runtime validator behavior based on that configuration.

The framework extends the RPC interface of invalidator such that
adversarial features can be adjusted on-the-fly while the invalidator is
running.

Co-authored-by: Trent Nelson <trent@solana.com>