adversary: Add feature to send shreds to a shred spammer (#33)
* Add feature

* Use feature set in broadcast and retransmit

---

Initial version was using iterators for filtering.  As more changes are
done later in this area, adding more variability to the packet
generation, an imperative style seems better.  It seems easier to
follow, compared to intertwined iterators.

In particular, `flat_map`s provide for multiple escape points, that are
not very noticeable, with the `?` applied to several layer of maps at a
time.

There is also an obvious performance optimization for the attack part:
when the attack target is not selected, we do not iterate over the
`shreds` slice at all.

Modified-by: Illia Bobyr <ilia.bobyr@solana.com>