pre-rebase-hook: Correctly forward to in-tree implementation (#132)
Previous version was just broken.  It is supposed to forward execution
to an implementation that is in the source tree, but it was looking for
the implementation in the wrong location.

`git rev-parse --git-dir` is a more robust way.  And we need `git` to
work for this script to make any sense.