transaction_generators: Reformat using iterators (#181)
I think these particular cases are easier to read when written using the
Iterator tooling, compared to more traditional for loops.  As a bonus,
most of the mutable variables are gone now.

Also move some of the state that does not change from one generator run
to another outside of the closure.  Most likely it does not affect the
generated code, but it can help the reader to understand what is
changing from one run to the next, and what is not.