adversary: Move attack configuration validation next to the generator (#183)
Configuration is tightly bound to the execution.  So it is much nicer
when they are defined next to each other.  In particular, they can now
share constants and other code with more ease.

Unfortunately, this refactoring requires a late binding mechanism for
the validation.  But the upside is that if we want to consolidate a
notion of an attack, we would need to implement something like this one
way or another.