get_program_ids: Replace `Vec` with `Option<[&Pubkey; N]>` (#321)
While the compiler is good at optimizing complex logic, it seems better
if we avoid explicit allocations when we do not really need them.  There
are only two uses of `get_program_ids_from_transaction()` - in one case
we check the first program ID and in the second with check the first
two.  Using `Option<[&Pubkey; N]>` better reflects our intention, at the
same time, removing an explicit allocation of a `Vec`.

Also use `.unwrap_or(false)` instead of `.unwrap_or_default()` for
booleans.  It is approximately the same length, but, I think, it is much
clearer that we want `false` when the value is not present.