Add transaction bench client (#339 + ...)
This change includes all of the following changes as well:

---

Author: kirill lykov <kirill.lykov@solana.com>
Date:   Tue May 28 20:20:39 2024 +0200

Add transaction bench client (#339)

New client application that sends transactions using TPU protocol.

First, this client creates accounts necessary for the transaction
generation.

Second, it generates transactions executing specified program and sends
them to the upcoming leaders.

Contrary to other clients, it doesn't use the `TpuClient` and the
`ConnectionCache` to interact with validators, but uses an alternative
client network layer implementation.

---

Author: Alex Peng <125332243+alpeng-jump@users.noreply.github.com>
Date:   Tue Jun 4 02:31:16 2024 -0500

transaction-bench: fix underflow in ensure_authority_balance (#341)

---

Author: kirill lykov <kirill.lykov@solana.com>
Date:   Fri Jul 26 10:02:39 2024 +0200

add 3 tests for client network component (#365)

This PR is the first in a series of PR introducing integration tests for
client network component.
It introduces the following changes:

* add first basic integration test

* add test_connection_denied_until_allowed

* add test for connection pruning

* retry connecting when pruned

* Introduce IoErrorWithPartialEq to check that QuicError is as expected

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Mon Aug 5 18:30:09 2024 +0200

Handle error with connection properly (#383)

When one of the connections fails to open, we stop the whole client.
It doesn't work on testnet because there most of the nodes drop TPU
traffic despite of saying through gossip that it is allowed (need to
ping foundation on that).

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Thu Aug 8 22:25:54 2024 +0200

Read/write for AccountsFile to use on testnet (#381)

Only the `transaction-bench/src/network/` and `transaction-bench/tests`
parts of this change was merge in.

Illia Bobyr: Migrated to quinn 0.11 and refactored the tests a bit.

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Sat Aug 17 15:10:19 2024 +0200

Handle instruction error in client-test-program (#411)

Due to upstream changes errors are not propagated properly anymore when
calling try_from_slice. This PR fixes the problem exactly the same way
as in #387

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Tue Aug 20 10:01:36 2024 +0200

Add structure to collect statistics about send txs erros (#367)

There is no way to check why certain packets have not been delivered.
By propagating this information to the caller, we can use it not only in
the client code but also in tests.
This change Introduces a structure which accumulates relevant counters
per node IP and returns this information to the client code.

Illia Bobyr: Migrated to quinn 0.11 and refactored a bit.

Refactored-by: Illia Bobyr <illia.bobyr@anza.xyz>

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Mon Aug 26 18:37:07 2024 +0200

Add rate limiting test to client (#368)

This PR introduces integration test to check that when rate-limiting has
happen we increment correct counters on client side.

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Tue Aug 27 16:34:57 2024 +0200

Improve stream sending (#372)

* stop using multistream connection

* add throttling test

* remove stream's finish call. It will be called when stream is dropped
  anyways, but if we call finish explicitly we will wait for ACK, so the
  behavior will be like if we had RW=1 tx. This change makes a x1000
  difference on a connection with high latency.

Illia Bobyr: Migrated to quinn 0.11, added cancellation in order to
support clean shutdown, and refactored the tests a bit.

Refactored-by: Illia Bobyr <illia.bobyr@anza.xyz>

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Thu Sep 5 18:54:01 2024 +0200

use testing_utilities (#427)

We use modified copy-pasted code from streamer to setup server/client in
integration tests for network component.

Later original code was updated to be reusable.

This PR removes old copy-pasted code and uses modified streamer's
testing_utilities code.

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Tue Sep 10 12:06:08 2024 +0200

Move leader updater to network and add documentation (#426)

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Tue Sep 10 16:30:34 2024 +0200

Add documentaiton to network code (#425)

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Tue Sep 10 20:58:04 2024 +0200

Don't ignore `test_no_host` (#435)

---

Author: kirill lykov <kirill.lykov@anza.xyz>
Date:   Wed Sep 11 20:34:09 2024 +0200

Address Illia's comments (#438)