pre-rebase-hook: Use `git branch --format`
As `git branch` can already output data in the format that we need, it
seems both more efficient and robust, if we avoid a transformation with
`sed`.  It also eliminates usage of `sed`, which is good, as MacOS uses
a version that is much more restrictive than Linux distributions.  So we
had several incompatibility issues, as all the testing happens on Linux.