Extract create_accounts function and add tests (#410)
Creating accounts on testnet is very unreliable process with numerous
errors.  To have a better implementation of the account creation, added
unit tests with mock RpcClient.  This allowed to simulate different
network scenarios.

In addition, as better than nothing approach, added commit to save
partial results in case if we failed to create requested number of
accounts. This is needed to get back funds from these accounts later.
Currently, we don't have a functionality to delete accounts and get back
their funds.