SQUASH Split max_size_accounts_tx out of max_accounts_tx (#454)
It seems simpler to have two distinct generator templates: one for
attacks that want to use the maximum possible number of accounts in a
transaction, and another one that uses the maximum possible number of
accounts of the maximum size.

These two templates have constraints that are quite different, and it
simplifies both the implementation and the interfaces of the templates.

SQUASH: This change should be incorporated into the following changes:

  Brennan <brennan.watt@anza.xyz>
  Tue Sep 3 21:28:31 2024 -0700
  constrain max accounts, shrink batch size (#419)

  Brennan <brennan.watt@anza.xyz>
  Thu Apr 18 18:10:43 2024 -0700
  read max accounts attack (#327)

  Illia Bobyr <illia.bobyr@solana.com>
  Mon Oct 23 12:21:02 2023 -0700
  Read and write lock attack (#196)