docs/upstream-sync.md: Latest instructions (#462)
Since I wrote the sync guide, I've realized a number of things and
actually tried some suggests that were marked as TODO.

While I've published a number of updates to the guide, an end-to-end
pass was still pending.  Now that we will have more people using the
guide on a regular basis, it makes sense to put as much of the things
I've learned into writing as possible.