Embed `solana-sbf-block-generator-stress-test` binary
When we run the compilation toolchain, there is a number of
dependencies that may break the compilation process.  In particular, we
see failures in our CI agents with the current version.

It is more conservative, for now, to just embed the program bytes
directly.  It reduces the complexity of the test execution.

Manual compilation is suboptimal, as the program binary might get out of
sync with the program source.

I have a draft change to run the compilation in the `build.rs` of the
`block-generator-stress-test-bin` package, but it is not working yet.