upstream sync: Avoid race condition of a parallel `master` update (#556)
git has a `--force-with-lease` option that makes non-fast-forward branch
updates safer by ensuring that the state of the branch that is
overwritten matches the expectation of the person overwriting the
branch.

This flag was added in git v1.8.5, released in 2013, and, though it has
been enhanced since then, it should be present in most git clients.