pre-rebase-hook: Run `git rev-list` without `--pretty`
When I was originally writing this code, I've used `git rev-list
--pretty=oneline` to get a list of changes, one per line.  But `git
rev-list` prints one line per change by default, as it outputs change
hashes.  Addition of `--pretty=oneline` increases `git rev-list`
execution time by about a factor of 3.

In two cases, `git rev-list` output is immediately passed into `wc -l`,
and the other case we just compare outputs of two `git rev-list`
commands to see if they differ.  I wonder if the latter can be done more
efficiently.

The slowness is mostly noticeable in the computation of the baseline, as
we repeat the `git rev-list` command for each of the `master/local/*`
tracking branch.  Of which we have 81 as of today.

For me, this change reduced baseline time calculation from 3408ms to
982ms.  It is still height for an interactive command and with more
`master/local/*` branches the time will keep growing linearly, but it is
a nice improvement.